Update nPose Menu.lsl
There was only 2 places toucherid would be used in a different way, touch event and link message receiver.  This branch separates toucherid in those two events and declares them locally rather than globally.  I saw one place where toucherid was not being assigned properly in the link message receiver and added a fix for that as well.

toucherid will be unique to the scope of each pass of the link message receiver.