Update nPose Slave.lsl
Doing the change this way creates a far smaller scope and I believe solves all the problems except the spam from the Adjuster script.

I will add the Adjuster script change in a moment to fix that issue also.  ((Gecko doesn't see a change when Adjusters are rezzed because it gets the real location from the chat line and doesn't see a difference in position/rotation.  When rezzing the adjusters in their final location which the adjuster doesn't recognize is causing the spam.))