Fix issues #60, #62, and part of #83
Added an option to allow quiet adjusters. Adjusters only report when PosDump is clicked when enabled. Issue #60
Added an option so that builder can pick to reference position/rotation to either the root prim or the prim that the slave is in.
Added custom seat names to adjuster reporting. Issue #62
Added NC name to PosDump reporting.
Cleaned up the code after changes.