added some more stuff and fixing some issues
- Added "Button Comments":
you may use /*YourComment*/ inside a button name. It will not be shown
in the dialog. It will be reported back in a dialog response. There are
two main scenarios where this may be usefull:
a) as a "poor man" button sorting option. Example:
BTN:/*010*/firstButton
BTN:/*020*/secondButton
BTN:/*030*/thirdButton
BTN:/*040*/fourthButton
b) to pass variables through the dialog script. Example: In change seat
menu plugin buttons are generated with the slotnumber as a comment:
/*0*/AnAvatarName
/*1*/AnotherAvatarName
/*3*/ASeatName
/*4*/ASeatNumber
So when the User presses a button we don't have to determine if the
button text is a Avatar Name a seat Name or whatever. We can simply
extract the slotnumber from the comment.
There are also no length restrictions for the button name anymore, so we
could even make a button like that:
/*6934889c-67c7-4b1d-9bbe-9cf84f1d12ad*/Leona (slmember1 Resident)

- Added "UTF for Buttons"
Currently we can't make nice checkboxes or anything like that, because
the character set of inventory items is very limited. Now you may use
/$Base64EncodedString$/ inside a button name.
Example:
BTN:/$4pyU$/ firstButton
will be rendered as "✔ firstButton".
Keep in mind that the length of the name of an inventory item is also
very limited (63Bytes)

- partly rewritten dialog generation

- prepared for user defined dialog prompt template

- prepared for user defined back/forward/backward buttons

- added a option for using backwards buttons

- added page display (Page: x/y)

- improved timeout handling

- improved memory usage by using str="";

- Bugfixing

- Beautifying